[CT-264] [CT-262] [Feature] Support BigQuery labels and hours_to_expiration configs for seeded models
