+++ tests/integration/simple_seed_test/models-bq/schema.yml
@@ -45,3 +45,10 @@ seeds:
     tests:
     - column_type:
         type: STRING
+
+- name: seed_configs
+  config:
+    hours_to_expiration: 2
+    labels:
+      contains_pii: 'yes'
+      contains_pie: 'no'
