+++ tests/integration/simple_seed_test/seeds-config/seed_configs.csv
@@ -1,0 +1,3 @@
+id,stuff
+1,a
+2,b
